Refactor calendar components and event handling
- Removed unused InteractionPlugin imports from Calendar and CalendarWeek components.
- Consolidated NewEvent component for both mobile and desktop views.
- Created a new CheckConnection utility for database connectivity checks.
- Added SuccessModal component for event creation confirmation.
- Deleted redundant NewEventDesktop and NewEventMobile components.
- Introduced new styles for the NewEvent component and its mobile variant.
- Updated ViewSchedule to utilize the new NewEvent component and improved responsiveness with useMediaQuery.
- Cleaned up UserScheduleCard styles and adjusted button styles for mobile.
