+++ src/components/CalendarWeek/index.tsx
@@ -1,14 +1,12 @@
 
 import FullCalendar from "@fullcalendar/react";
-import InteractionPlugin from "@fullcalendar/interaction";
 import dayGridPlugin from "@fullcalendar/daygrid";
 import timeGridPlugin from "@fullcalendar/timegrid";
 import "./style.css";
-import { use, useEffect, useState } from "react";
-import NewEvent from "../NewEvent/NewEventDesktop/NewEventDesktop";
+import {  useEffect, useState } from "react";
 import { checkDebugConnection } from "../CheckConnection/CheckConnection";
 import { getEvents } from "../../constants/calendar";
-import type { CalendarDTO } from "../../models/calendar";
+import NewEvent from "../NewEvent";
 export default function CalendarWeek() {
