+++ src/components/CheckConnection/CheckConnection.ts
