+++ src/components/Modal/SucessModal.tsx
@@ -1,16 +1,17 @@
 import "./sucessModalMobile.css";
+import "./sucessModal.css";
 
-export default function SuccessModalMobile({closeThen, title, content }: { closeThen: React.Dispatch<React.SetStateAction<boolean>>; title?: string; content?: string }) {
+export default function SuccessModal({ isMobile, closeThen, title, content }: { isMobile: boolean; closeThen: React.Dispatch<React.SetStateAction<boolean>>; title?: string; content?: string }) {
 
     function handleCloseModal() {
-       
+
         closeThen(false);
     }
 
     return (
         <>
             <div className="overlay"></div>
-            <div className="modal-event-created-mobile">
+            <div className={`modal-event-created${isMobile ? "-mobile" : ""}`}>
                 <h2>{title || "Evento criado com sucesso!"}</h2>
                 <p className="content-modal">{content || "Seu evento foi criado com sucesso."}</p>
                 <svg width="52" height="51" viewBox="0 0 52 51" fill="none" xmlns="http://www.w3.org/2000/svg">
