+++ src/routes/Schedule/mobile.css
@@ -66,8 +66,3 @@
 
     }
 }
-
-
-.btn-sched-mobile {
-    padding: 15px 10px !important;
-}
